rename package main -> boardreader
